create ClassConstFetchScopeNodeVisitor
